fix signed conversion warning
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/Performance@16740 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/AsyncFramework/JobScheduler.cpp
@@ -439,7 +439,8 @@ void CJobScheduler::WaitForSomeJobs()
 {
     {
         CCriticalSectionLock lock (mutex);
-        if (InterlockedIncrement (&waitingThreads) < threads.runningCount)
+        if (  static_cast<size_t>(InterlockedIncrement (&waitingThreads))
+            < threads.runningCount)
         {
             // there are enough running job threads left
             // -> wait for one of them to run idle *or*
